init file move
