+++ packages/mongodb-chatbot-server/src/processors/index.ts
@@ -1,14 +1,10 @@
 export * from "./FilterPreviousMessages";
-export * from "./FindContentFunc";
-export * from "./FindVerifiedAnswerFunc";
 export * from "./GenerateUserPromptFunc";
 export * from "./MakeReferenceLinksFunc";
 export * from "./QueryPreprocessorFunc";
-export * from "./SearchBooster";
+export * from "mongodb-rag-core/src/SearchBooster";
 export * from "./filterOnlySystemPrompt";
-export * from "./makeBoostOnAtlasSearchFilter";
-export * from "./makeDefaultFindContent";
-export * from "./makeDefaultFindVerifiedAnswer";
+export * from "mongodb-rag-core/src/makeBoostOnAtlasSearchFilter";
 export * from "./makeDefaultReferenceLinks";
 export * from "./makeFilterNPreviousMessages";
 export * from "./makeRagGenerateUserPrompt";
