+++ packages/mongodb-rag-core/src/index.ts
@@ -1,7 +1,11 @@
 export * from "./CoreEnvVars";
 export * from "./DatabaseConnection";
+export * from "./DefaultFindContent";
+export * from "./DefaultFindVerifiedAnswer";
 export * from "./EmbeddedContent";
 export * from "./Embedder";
+export * from "./FindContentFunc";
+export * from "./FindVerifiedAnswerFunc";
 export * from "./LangChainEmbedder";
 export * from "./MongoDbEmbeddedContentStore";
 export * from "./MongoDbVerifiedAnswerStore";
@@ -12,6 +16,7 @@ export * from "./PageFormat";
 export * from "./TypeChatJsonTranslateFunc";
 export * from "./VectorStore";
 export * from "./VerifiedAnswer";
+export * from "./SearchBooster";
 export * from "./arrayFilters";
 export * from "./assertEnvVars";
 export * from "./extractFrontMatter";
